requirements.txt
assemblyai[extras]
google-generativeai
cartesia
pyaudio
python-dotenv
